renamed component to distinguish Entity vs Pages
